feat(fill-blanks): extract droppable; adjust styles; clean code
